refactor: remove duplicated file
