+++ packages/trader/src/App/Components/Elements/ContractDrawer/contract-drawer-card.jsx
@@ -2,8 +2,14 @@ import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { DesktopWrapper, MobileWrapper, Collapsible, ContractCard, useHover } from '@deriv/components';
-import { isCryptoContract, isDesktop, getEndTime, getSymbolDisplayName } from '@deriv/shared';
-import { getCardLabels, getContractTypeDisplay } from 'Constants/contract';
+import {
+    getEndTime,
+    getSymbolDisplayName,
+    getCardLabels,
+    getContractTypeDisplay,
+    isCryptoContract,
+    isDesktop,
+} from '@deriv/shared';
 import { getMarketInformation } from 'Utils/Helpers/market-underlying';
 import { SwipeableContractDrawer } from './swipeable-components.jsx';
 import MarketClosedContractOverlay from './market-closed-contract-overlay.jsx';
