+++ packages/trader/src/App/Containers/ProgressSliderStream/progress-slider-stream.jsx
@@ -1,8 +1,7 @@
 import PropTypes from 'prop-types';
 import React from 'react';
 import { ProgressSlider } from '@deriv/components';
-import { getCurrentTick } from '@deriv/shared';
-import { getCardLabels } from 'Constants/contract';
+import { getCardLabels, getCurrentTick } from '@deriv/shared';
 import { observer, useStore } from '@deriv/stores';
 
 const ProgressSliderStream = observer(({ contract_info }) => {
