+++ packages/trader/src/Constants/index.js
@@ -1,2 +1,1 @@
-export * from './contract';
 export * from './ui';
