+++ packages/trader/src/Modules/Contract/Components/Digits/digits.jsx
@@ -3,10 +3,9 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { toJS } from 'mobx';
 import { DesktopWrapper, MobileWrapper, Popover, Text } from '@deriv/components';
-import { isMobile, useIsMounted, isContractElapsed } from '@deriv/shared';
+import { getMarketNamesMap, isMobile, useIsMounted, isContractElapsed } from '@deriv/shared';
 import { localize, Localize } from '@deriv/translations';
 import { Bounce, SlideIn } from 'App/Components/Animations';
-import { getMarketNamesMap } from '../../../../Constants';
 import { DigitSpot, LastDigitPrediction } from '../LastDigitPrediction';
 import 'Sass/app/modules/contract/digits.scss';
